+++ src/DotNetWorker.Core/FunctionMetadata/IFunctionMetadataTransformer.cs
@@ -1,7 +1,7 @@
-// Copyright (c) .NET Foundation. All rights reserved.
+﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
-using System.Collections.Immutable;
+using System.Collections.Generic;
 
 namespace Microsoft.Azure.Functions.Worker.Core.FunctionMetadata;
 
@@ -12,11 +12,13 @@ namespace Microsoft.Azure.Functions.Worker.Core.FunctionMetadata;
 public interface IFunctionMetadataTransformer
 {
     /// <summary>
+    /// Gets the name of the transformer.
+    /// </summary>
+    public string Name { get; }
+
+    /// <summary>
     /// Transforms the provided collection of <see cref="IFunctionMetadata"/> instances.
     /// </summary>
-    /// <param name="original">The original collection of function metadata.</param>
-    /// <returns>
-    /// An <see cref="ImmutableArray{IFunctionMetadata}"/> containing the transformed function metadata.
-    /// </returns>
-    ImmutableArray<IFunctionMetadata> Transform(ImmutableArray<IFunctionMetadata> original);
+    /// <param name="original">The original collection of function metadata. This collection may be modified.</param>
+    void Transform(IList<IFunctionMetadata> original);
 }